requirements.txt
aiogram==3.*
python-dotenv==1.*
asyncpg==0.29.*